### 2021-11-02 #3880 @whedon
[PRE REVIEW]: Sinaps: A Python library to simulate voltage dynamic and ionic electrodiffusion in neurons
**Submitting author:** @cgurrr (<a href="http://orcid.org/0000-0002-1128-1069">Claire Guerrier</a>)
**Repository:** <a href="https://github.com/ngltr/sinaps" target ="_blank">https://github.com/ngltr/sinaps</a>
**Version:** 0.2
