+++ README.md
@@ -1,3 +1,5 @@
+[![status](https://joss.theoj.org/papers/d47fbf887ce21cf686a56bac1b36817e/status.svg)](https://joss.theoj.org/papers/d47fbf887ce21cf686a56bac1b36817e)
+
 # SiNAPS
 
 `Sinaps` is a Python package providing a fast, flexible and expressive tool to model signal propagation and ionic electrodiffusion in neurons. It is an efficient framework to build computational models of voltage and calcium dynamics. It is based on the Cable theory for voltage propagation, and the Nernst-Planck equation for the electrodiffusion of ions.
