+++ README.md
@@ -1,4 +1,5 @@
 [![status](https://joss.theoj.org/papers/d47fbf887ce21cf686a56bac1b36817e/status.svg)](https://joss.theoj.org/papers/d47fbf887ce21cf686a56bac1b36817e)
+[![status](https://img.shields.io/pypi/v/sinaps?color=green)](https://pypi.org/project/SiNAPS/)
 
 # SiNAPS
 
