### 2021-11-03 #3880 @meg-simula
Sure, right up my alley.
