### 2021-11-06 7c9952f
unit correction in section display
