### 2021-11-07 8d4362d
add diffusion coefficient error
