+++ tests/unit/test_api.py
@@ -4,14 +4,12 @@ from sinaps import Neuron
 
 def test_neuron_repr():
     nrn = Neuron()
-    assert(isinstance(nrn.__repr__(), str))
-    assert(len(nrn.__repr__())) > 0
+    assert isinstance(nrn.__repr__(), str)
+    assert (len(nrn.__repr__())) > 0
 
 
+# %%
 
 
-
-#%%
-
-if __name__ == '__main__':
-   test_neuron_repr() 
+if __name__ == "__main__":
+    test_neuron_repr()
