### 2021-11-11 8f17178
Add voltage source to the model
