+++ tests/integration/test_simu.py
@@ -1,4 +1,5 @@
 import sinaps as sn
+import numpy as np
 
 
 def test_HH():
@@ -13,3 +14,14 @@ def test_HH():
     sim.run((0, 20))
 
     assert sim.V.max().min() > 110
+
+
+def test_voltage_source():
+    nrn = sn.Neuron()
+    sec = sn.Section(L=1000)
+    sec.add_voltage_source(sn.VoltageSource(lambda t: np.sin(t)))
+    nrn.add_section(sec, 0, 1)
+    sim = sn.Simulation(nrn, 10)
+    sim.run((0, 10))
+
+    assert sim.V.iloc[:, 1].max() > 0.6
