+++ README.md
@@ -1,10 +1,13 @@
 [![status](https://joss.theoj.org/papers/d47fbf887ce21cf686a56bac1b36817e/status.svg)](https://joss.theoj.org/papers/d47fbf887ce21cf686a56bac1b36817e)
 [![status](https://img.shields.io/pypi/v/sinaps?color=green)](https://pypi.org/project/SiNAPS/)
-![status](https://github.com/ngltr/sinaps/workflows/pytest/badge.svg)
+[![status](https://github.com/ngltr/sinaps/workflows/pytest/badge.svg)](https://github.com/ngltr/sinaps/actions/workflows/python-app.yml)
+
 
 # SiNAPS
 
 `Sinaps` is a Python package providing a fast, flexible and expressive tool to model signal propagation and ionic electrodiffusion in neurons. It is an efficient framework to build computational models of voltage and calcium dynamics. It is based on the Cable theory for voltage propagation, and the Nernst-Planck equation for the electrodiffusion of ions.
+
+![image](docs/_static/bAP.gif)
 
 ## Documentation
 
