### 2021-11-30 #3880 @meg-simula
@cgurrr Thanks for this submission and my apologies for the delay in handling.

If you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
