### 2021-12-20 #4012 @whedon
[REVIEW]: Sinaps: A Python library to simulate voltage dynamic and ionic electrodiffusion in neurons
**Submitting author:** @cgurrr (<a href="http://orcid.org/0000-0002-1128-1069">Claire Guerrier</a>)
**Repository:** https://github.com/ngltr/sinaps
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.2
