### 2021-12-21 ngltr/sinaps#2 @martejulie
Empty test file
Hi! Very nice work! :)
I was looking through the test files and noticed that the
tests/unit folder contains an empty file, test_channel.
