### 2021-12-21 ngltr/sinaps#3 @martejulie
Failed doctest
Hi! :)
I get the following message when I try to run the doctests in model.py:
```
**********************************************************************
File "model.py", line 181, in __main__.Neuron.add_species
Failed example:
    nrn.add_species(Species.Ca, C0=1)
Exception raised:
    Traceback (most recent call last):
      File "/Applications/Xcode.app/Contents/Developer/Library/Frameworks/Python3.framework/Versions/3.8/lib/python3.8/doctest.py", line 1329, in __run
        exec(compile(example.source, filename, "single",
      File "<doctest __main__.Neuron.add_species[2]>", line 1, in <module>
        nrn.add_species(Species.Ca, C0=1)
      File "model.py", line 207, in add_species
        raise InitialConcentrationError(sp)  # TODO
    InitialConcentrationError: No initial concentration defined for the ion Ca2+
                   define it using [section].C0=dict([ion]=[initial_concentration])

**********************************************************************
File "model.py", line 182, in __main__.Neuron.add_species
Failed example:
    nrn.sections
Expected nothing
Got:
    {Section(C0={<Species Ca2+>: 1}, C_m=1, D={}, L=100, R_l=150, V0=0, a=1, dx=None, name='Section0002'): (0, 1)}
**********************************************************************
1 items had failures:
   2 of   4 in __main__.Neuron.add_species
***Test Failed*** 2 failures.
```
