### 2021-12-21 ngltr/sinaps#4 @martejulie
Fieldplots
I’m running the plotting notebook from the documentation and have a couple of questions/issues regarding the field plots.

1. How are the positions on the y-axis mapped to the neuronal structure/sections? I think it would be of great value to clarify this in the documentation as well as in Figure 1 of the paper.

2. The field plots are missing color bars.
