### 2021-12-22 ngltr/sinaps#6 @martejulie
Readme
It would be great if the animation in README came with a figure caption and a colorbar.
