### 2021-12-22 ngltr/sinaps#7 @martejulie
Model assumptions
It would be helpful for the user if the main assumptions of the model were explicitly stated in the documentation, i.e., zero extracellular potential and constant reversal potentials.
