### 2021-12-22 ngltr/sinaps#8 @martejulie
Model equations
The documentation states that Sinaps is based on the Cable theory for voltage propagation, and the Nernst-Planck equation for the electrodiffusion of ions, but I miss an explicit formulation of the equations. The equations are given in the docstrings of _ode_function and _ode_diff_function in the Simulation class, but they don’t come with a full description of the parameters/variables. It would be good to add this description to the docstrings and great if it was also added to the documentation.
