### 2021-12-22 ngltr/sinaps#9 @martejulie
HH equations
For the Hodgkin-Huxley equations, why are you solving only one ODE for the state variables?
