### 2021-12-22 #4012 @martejulie
Hi! Very nice work!

I support the publication of this software with minor revisions.

Here comes a few questions and some (minor) suggestions for improvement.

**Functionality**
- Please see issue [#4](https://github.com/ngltr/sinaps/issues/4) in the source code repo.

**Documentation**
- Please see issue [#6](https://github.com/ngltr/sinaps/issues/6), [#7](https://github.com/ngltr/sinaps/issues/7), [#8](https://github.com/ngltr/sinaps/issues/8), and [#9](https://github.com/ngltr/sinaps/issues/9) in the source code repo.

**Automated tests**
- Please see issue [#2](https://github.com/ngltr/sinaps/issues/2) and [#3](https://github.com/ngltr/sinaps/issues/3) in the source code repo.

**Quality of writing**
- Please see my [pull request](https://github.com/ngltr/sinaps/pull/5) for some minor suggestions.

**State of the field/References**
- Numerous papers have been published on the modeling of voltage dynamics and ion concentration dynamics in neurons, see [Sætra et al. 2020](https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1007661) and the references therein. The novelty of this work is the fast computation of intracellular electrodiffusion of ions for complex neuronal geometries and the development of a Python package that can be used to build custom set-ups. I think it would be appropriate to add a short paragraph (two-three sentences) about this in the paper. A key reference, which I think should be cited, is the paper of Solbrå et al. 2018, [A Kirchhoff-Nernst-Planck framework for modeling large scale extracellular electrodiffusion surrounding morphologically detailed neurons](https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1006510), and their simulator [KNPsim](https://github.com/CINPLA/KNPsim).
