### 2021-12-25 61b0db0
Add colorbar to field plot (fix part 2 of #4)
