### 2021-12-31 badb4ee
Fix #3 (correct doc for add_species)
