### 2022-01-01 4f9f46f
Add automated test for channeles - Fix #2
