### 2022-01-02 7b04f9d
Add neuron representation and correlation with field plot - Fix #4
