+++ src/sinaps/gui/graph.py
@@ -4,6 +4,7 @@ import hvplot.pandas
 import pandas as pd
 import holoviews as hv
 from holoviews.operation.datashader import rasterize
+import networkx as nx
 
 
 class Plotter:
@@ -41,19 +42,25 @@ class SimuView:
     def __init__(self, simu):
         self.simu = simu
 
-    def field(self, df, time=None, res=None):  # TODO make it a static method
+    def field_plot(self, df, title, time=None, res=None, neuron=True, dynamic=True):
         # res : resolution : #um
+        nrn = self.simu.N
+        #### field plot
         if time is not None:
             df = df[slice(*time)]
         dfi = interpolate.interp1d(
-            self.simu.N.indexV_flat(), df.values, fill_value="extrapolate"
+            nrn.indexV_flat(), df.values, fill_value="extrapolate"
         )
         if res is None:
             res = max(max(self.simu.N.indexV_flat()) / 1000, 1)
         y = np.arange(0, max(self.simu.N.indexV_flat()) + res, res)
-        return rasterize(hv.QuadMesh((df.index.values, y, dfi(y).T))).opts(
-            xlabel="Time (ms)",
-            ylabel="Position (μm)",
+
+        p_field = rasterize(
+            hv.QuadMesh(
+                (df.index.values, y, dfi(y).T), ["Time (ms)", "Position (μm)"], title
+            )
+        )
+        p_field.opts(
             width=600,
             height=600,
             tools=["hover"],
@@ -61,18 +68,132 @@ class SimuView:
             colorbar=True,
         )
 
+        if not neuron:
+            return p_field
+        else:
+            #### Neuron representation
+            g_2 = nx.Graph()
+            edges = nrn.sections.values()
+            ej = nrn.__traversal_source__
+            branch = 0
+            x = 0
+            line = []
+            path = [0]
+            for e in edges:
+                if e[0] != ej:
+                    line += [
+                        {
+                            "Position (μm)": path,
+                            "h": [0] * len(path),
+                            "branch": branch % 20,
+                        }
+                    ]
+                    branch += 1
+                    path = [x]
+
+                sec = nrn.graph.edges[e]["section"]
+                g_2.add_edge(e[0], e[1], branch=branch, x=x)
+                ej = e[1]
+                path.append(x + sec.L)
+                x += sec.L
+            line += [
+                {"Position (μm)": path, "h": [0] * len(path), "branch": branch % 20}
+            ]
+
+            p_graph = hv.Graph.from_networkx(g_2, nrn.plot.layout)
+            p_line = hv.Path(line, ["h", "Position (μm)"], vdims="branch")
+
+            p_graph.opts(
+                width=300,
+                height=600,
+                padding=0.1,
+                node_size=0,
+                xaxis=None,
+                yaxis=None,
+                edge_line_width=2,
+                edge_color=hv.dim("branch") % 20,
+                edge_cmap="Category20b",
+                inspection_policy="edges",
+                edge_hover_line_color="green",
+            )
+            p_line.opts(
+                color="branch",
+                cmap="Category20b",
+                line_width=20,
+                xaxis=None,
+                yaxis=None,
+                height=600,
+                width=30,
+            )
+
+            if not dynamic:
+                return p_graph + p_line + p_field
+            else:
+
+                #### Dynamic selection
+
+                Branch = hv.streams.Stream.define("branch", branch=None)
+                branch_stream = Branch()
+                # Edge selection
+                dmap1 = hv.DynamicMap(
+                    lambda branch: p_graph.select(branch=branch),
+                    streams=[branch_stream],
+                )
+
+                # Field selection
+                t0, t1 = df.index[0], df.index[-1]
+
[…]
