### 2022-01-02 ngltr/sinaps#4 @ngltr
The field plot is modified to show also the neuron structure and a colormap represents the mapping between the y-axis and the neuronal structure
![image](https://user-images.githubusercontent.com/49352943/147866855-7f24ecf4-1b55-44c1-bafe-1d1559c74b0b.png)
