+++ src/sinaps/core/model.py
@@ -823,7 +823,7 @@ class Section(param.Parameterized):
         return np.diff(self.xb) * self._param_array(self.a) * 2 * PI  # [um2]
 
     def _r_l_array(self):
-        """Return the longitunal resistance between each nodes
+        """Return the longitudinal resistance between each nodes
         init_sim(dx) must have been previously called
         """
         # np.diff(self.x) = distance between centers of compartiment
@@ -834,7 +834,7 @@ class Section(param.Parameterized):
         )  # [MΩ]
 
     def _r_l_end(self):
-        """Return the longitunal resistance between the start/end of the section
+        """Return the longitudinal resistance between the start/end of the section
         and the first/last node
         init_sim(dx) must have been previously called
         """
