### 2022-01-10 5c7c515
Add model description and assumptions in the documentation - Fix #8 Fix #7
