+++ docs/model.md
@@ -1,0 +1,257 @@
+Model Description
+-----------------
+
+To model voltage propagation and ionic electrodiffusion in the *Sinaps*
+python package, we consider the Cable equation [1] for voltage
+dynamic, coupled to the Nernst-Planck equation for ionic dynamic.\
+The Cable equation makes an analogy between the dendritic tree, and a
+electric wire composed of RC circuits:
+
+$$c \frac{\partial  V}{\partial t} = I^G + \frac{1}{R} \frac{\partial^2 V}{\partial x^2}.$$
+
+In this model, the axonal membrane acts as a capacitor (capacitance
+$c$), in parallel with voltage-dependent conductors accounting for ionic
+channels (currents $I^G$). The voltage propagation is represented by a
+resistive current: the current transported by electrons in the solid
+structure of the wire follows Ohm's law (current $I = \frac{U}{R}$). This model assumes a zero extracellular potential. There are a priori no restrictions on membrane currents.\
+To infer calcium dynamics from voltage propagation, we couple the
+previous equation to the Nernst-Planck equation, for ionic
+electrodiffusion. This equation describes the motion of charged species
+within an electrolyte [2], and is build from the first principle
+of mechanics, stating the conservation of ions. We add to this equation
+the flux of ions coming from the channels, as well as the variation of
+concentration due to the eventual chemical reactions. We get:
+
+$$\label{NP}
+\frac{\partial c^i}{\partial t} = - \nabla \cdot j^i + j^{i,\text{m}} + \gamma^{i,\text{chem}} \;\; \forall i \in \text{\{ions\}}$$
+
+The flux of charge per surface area $j_i$ \[mol.s$^{-1}$.m$^{-2}$\],
+sums the diffusion flux of ionic species derived from Fick's law
+$j_{diff}^i = - D_i \nabla c_i$, and an electric flux representing the
+motion of ions due to the electric field:
+$j_{\text{elec}}^i = - \frac{D_i z_i e c_i}{k_b T} \nabla V$
+[2,3]. $\displaystyle \frac{D_i z_i e c_i}{k_b T}$ is the
+Einstein relation for the mobility of ion species: $D_i$ is the
+diffusion coefficient of species $i$, $z_i$ the valence of species $i$,
+$k_B T$ the thermal energy and $e$ is the electric charge [3]. We
+get:
+
+$$j^i = j_{diff}^i + j_{\text{elec}}^i = - D_i \nabla c^i  - \frac{D^i z^i e c^i}{k_b T} \nabla V.$$
+
+$j^{i, \text{m}}$ represents the lineic flux of ions entering from the
+channels \[mol.s$^{-1}$.m$^{-1}$\]. $\gamma^{i,\text{chem}}$ represents
+the variations of concentration due to the eventual chemical reactions.
+
+Notations
+---------
+
+In one branch going between $\tilde{x}_0 \leq x \leq \tilde{x}_L$, we
+consider different positions
+$\tilde{x}_0 < \tilde{x}_1<...< \tilde{x}_n <...< \tilde{x}_N < \tilde{x}_{N+1} = \tilde{x}_L$.
+We then define $x_n$ as the middle point between $\tilde{x}_n$ and
[…]
