### 2022-01-10 50adf7d
Correct initialization formultiple variable channels
