### 2022-01-13 ngltr/sinaps#9 @cgurrr
Hi,
The Hodgkin-Huxley system of equations we implemented is a simplification of the classical HH, with m = m_{\infty}, and h = 0.89 - 1.1 n (see Mathematical Physiology, Keener J. and Sneyd J., chap.5).
We will now change this implementation into Hodgkin-Huxley_red, and implement a full version similar to the system of equations in the 1952 paper (A QUANTITATIVE DESCRIPTION OF MEMBRANE
CURRENT AND ITS APPLICATION TO CONDUCTION
AND EXCITATION IN NERVE, Hodgkin, Huxley, J Physiol 1952)
