### 2022-01-13 0480512
Implement full version of HH channels - Fix #9
