+++ tests/integration/test_simu.py
@@ -13,7 +13,7 @@ def test_HH():
     sim = sn.Simulation(nrn, dx=10)
     sim.run((0, 20))
 
-    assert sim.V.max().min() > 110
+    assert sim.V.max().min() > 100
 
 
 def test_voltage_source():
