### 2022-01-14 473a37c
Update paper.md and references
