+++ paper.bib
@@ -1,11 +1,24 @@
 
+@article{Saetra:2020,
+    doi = {10.1371/journal.pcbi.1007661},
+    author = {Sætra, Marte J. AND Einevoll, Gaute T. AND Halnes, Geir},
+    journal = {PLOS Computational Biology},
+    title = {An electrodiffusive, ion conserving Pinsky-Rinzel model with homeostatic mechanisms},
+    year = {2020},
+    month = {04},
+    volume = {16},
+    pages = {1-36},
+    number = {4}
+
+}
 
 @article{Sakaki:2020,
         author = {Sakaki K and Podgorski K and Delazzizo Toth T and Coleman P and Haas K},
         title = {Comprehensive Imaging of Sensory-Evoked Activity of Entire Neurons Within the Awake Developing Brain Using Ultrafast AOD-Based Random-Access Two-Photon Microscopy},
         journal = {Front in Neural Circuit},
         volume = {14},
-        year = {2020}
+        year = {2020},
+	doi = {10.3389/fncir.2020.00033}
 }
 
 @article{Basak:2018,
@@ -14,7 +27,21 @@
         journal = {PLoS Comp Bio},
         volume = {14},
         pages = {e1006485},
-        year = {2018}
+        year = {2018},
+	doi = {10.1371/journal.pcbi.1006485}
+}
+
+@article{Solbra:2018,
+    doi = {10.1371/journal.pcbi.1006510},
+    author = {Solbrå, Andreas AND Bergersen, Aslak Wigdahl AND van den Brink, Jonas AND Malthe-Sørenssen, Anders AND Einevoll, Gaute T. AND Halnes, Geir},
+    journal = {PLOS Computational Biology},
+    title = {A Kirchhoff-Nernst-Planck framework for modeling large scale extracellular electrodiffusion surrounding morphologically detailed neurons},
+    year = {2018},
+    month = {10},
+    volume = {14},
+    pages = {1-26},
+    number = {10}
+
 }
 
 @Book{Carnevale:2006,
@@ -22,7 +49,8 @@
 	title = {The NEURON Book},
 	publisher = {Cambridge University Press},
 	year = {2006},
-	address = {Cambridge, UK}
+	address = {Cambridge, UK},
+	doi = {10.1017/CBO9780511541612}
 }
 
 
@@ -32,7 +60,8 @@
         journal = {Annu Rev Neurosci},
         volume = {28},
         pages = {503-532},
-        year = {2005}
+        year = {2005},
+	doi = {10.1146/annurev.neuro.28.061604.135703}
 }
 
 
@@ -42,5 +71,6 @@
 	title = {The Book of GENESIS: Exploring Realistic Neural Models with the GEneral NEural SImulation System, Second edition},
 	publisher = {Springer-Verlag},
 	year = {1998},
-	address = {New York}
+	address = {New York},
+	doi = {10.1086/419236}
 }
