+++ paper.md
@@ -27,7 +27,7 @@ bibliography: paper.bib
 How do neuronal dendrites collect, process and transmit information? What is the role of neurons'
 specific geometry on neuronal activity? The old postulate that dendrites serve mainly to connect
 neurons and to convey information, with no specific role in cognitive processes, is currently
-challenged by the emergence of novel experimental techniques [@London:2005 ; @Basak:2018]. Hence, the
+challenged by the emergence of novel experimental techniques [@London:2005 ; @Basak:2018]. Hence, the 
 role of the dendritic tree in transforming synaptic input into neuronal output is now a leading
 question in developmental neuroscience. In particular, using genetically-encoded Ca$^{2+}$-
 indicators, state-of-the-art techniques have been developed to track the calcium dynamics within
@@ -44,9 +44,11 @@ Here, we present `Sinaps`, which is an easy-to-use Python library to simulate vo
 
 # Statement of need
 
-The `Sinaps` library is based on the Cable equation for voltage propagation, coupled to the Nernst-Planck equation for ionic electrodiffusion. The library has been designed for neuroscience laboratories using both an experimental and a modeling approach. It includes the code to simulate voltage dynamics and ionic electrodiffusion, Hodgkin-Huxley type membrane channels, and chemical reactions. Templates to code custom reaction-diffusion mechanisms, as well as specific membrane channels are provided. We also provide the possibility to load a full morphometric geometry from data following [neuromopho.org](http://neuromorpho.org/) file type (swc file) \autoref{fig:volt}.
+Numerous softwares has been designed to realize simulations of voltage dynamics and ion concentration in neurons [@Bower:1998 ; @Carnevale:2006]. Most of those softwares are using the Cable theory model, and are designed toward neuronal networks simulation. Models for ion concentration dynamics in neurons, based on the Nernst-Planck model coupled to specific equations describing the potential, are also available [@Saetra:2020 ; @Solbra:2018]. 
 
-Numerous softwares has been designed to realize simulations of voltage propagation in neurons [@Bower:1998 ; @Carnevale:2006]. While most of those softwares are using the simple Cable theory model, and are designed toward neuronal networks simulation, our Python library is designed to realize fast simulation of both voltage and ionic dynamics, taking into account electrodiffusion of ions at a fine spatial scale. We also choose to realize the code in Python, which has the advantage of having a code fully transparent with easy access to all the variables. The class structure renders the code easily editable. Hence, our library provides an easy way to simulate voltage and ionic dynamics, at the spacial scale reached by morphometric techniques, and at a temporal scales not yet available for \textit{in vivo} imaging of the full neuronal scale.
+Among these models, the `Sinaps` library is designed to realize the fast computation of voltage propagation and ionic electrodiffusion in complex neuronal geometries, at a fine spatial scale. It is based on the Cable equation for voltage propagation, coupled to a one-dimensional Nernst-Planck equation for ionic electrodiffusion. The library can be used to build custom set-ups, with easy access to all the variables. Finally, the class structure renders the code easily editable.
+
+The library has been designed for neuroscience laboratories using both an experimental and a modeling approach. It includes the code to simulate voltage dynamics and ionic electrodiffusion, Hodgkin-Huxley type membrane channels, and chemical reactions. Templates to code custom reaction-diffusion mechanisms, as well as specific membrane channels are provided. We also provide the possibility to load a full morphometric geometry from data following [neuromopho.org](http://neuromorpho.org/) file type (swc file) \autoref{fig:volt}. Hence, our library provides an easy way to simulate voltage and ionic dynamics, at the spacial scale reached by morphometric techniques, and at a temporal scales not yet available for \textit{in vivo} imaging of the full neuronal scale.
 
 # Acknowledgements
 
