### 2022-01-17 5f87f81
Complete documentation on field plot ( see #4 )
