### 2022-01-17 8531f4a
Complet documentation on field plot ( see #4 )
