### 2022-01-17 f1b49cb
Rename HeavysideCurrent to PulseCurrent
