+++ tests/unit/test_channels.py
@@ -10,7 +10,7 @@ CHANNEL_LIST = [
     ch.ConstantCurrent(5),
     ch.AMPAR(1),
     ch.NMDAR(1),
-    ch.HeavysideCurrent(5, 1, 2),
+    ch.PulseCurrent(5, 1, 2),
     ch.Hodgkin_Huxley(),
     ch.Hodgkin_Huxley_red(),
     ch.Hodgkin_Huxley_Ca(),
@@ -22,11 +22,10 @@ CHANNEL_LIST = [
 def neuron_with_channel():
     sec = Section()
     sec.add_channel(ch.ConstantCurrent(5))
-    nrn = Neuron({sec:(0, 1)})
+    nrn = Neuron({sec: (0, 1)})
     return nrn
 
 
 def test_channels_I(neuron_with_channel):
     sim = Simulation(neuron_with_channel, dx=10)
     sim.run((0, 5))
-
