+++ tests/integration/test_simu.py
@@ -6,7 +6,7 @@ def test_HH():
     nrn = sn.Neuron()
     sec = sn.Section()
     sec.add_channel(sn.channels.Hodgkin_Huxley())
-    sec.add_channel(sn.channels.HeavysideCurrent(100, 2, 4), 0)
+    sec.add_channel(sn.channels.PulseCurrent(100, 2, 4), 0)
     nrn.add_section(sec, 0, 1)
     assert nrn[0] == sec
     assert nrn[1] == sec
