### 2022-01-20 #4012 @martejulie
Hi, @cgurrr!

Thank you! Very nice to see that you have followed all my suggestions. I have already checked off a few more marks on the review checklist.

It would be great if you could create a new pip release of the current version of the code and update the documentation. (I noticed that 1) the HeavysideCurrent function is still used in the Plotting notebook (on the web page), and 2) the API references are not up to date with the new Hodgkin-Huxley implementation. There might be more.)

This would make it easier to review the documentation and the functionality of the code.
