### 2022-01-20 ngltr/sinaps#6 @martejulie
@ngltr Please update the documentation frontpage accordingly.

(Btw, the bullet points below Model Description in the site map (Notations-Bibliography) should be sub-points. To fix this is of course not a requirement for publication, but the site map will give a better overview if you do.)
