+++ paper.bib
@@ -1,7 +1,7 @@
 
 @article{Saetra:2020,
     doi = {10.1371/journal.pcbi.1007661},
-    author = {Sætra, Marte J. AND Einevoll, Gaute T. AND Halnes, Geir},
+    author = {Sætra MJ and Einevoll GT and Halnes G},
     journal = {PLOS Computational Biology},
     title = {An electrodiffusive, ion conserving Pinsky-Rinzel model with homeostatic mechanisms},
     year = {2020},
@@ -33,7 +33,7 @@
 
 @article{Solbra:2018,
     doi = {10.1371/journal.pcbi.1006510},
-    author = {Solbrå, Andreas AND Bergersen, Aslak Wigdahl AND van den Brink, Jonas AND Malthe-Sørenssen, Anders AND Einevoll, Gaute T. AND Halnes, Geir},
+    author = {Solbrå A and Bergersen AW and van den Brink J and Malthe-Sørenssen A and Einevoll GT and Halnes G},
     journal = {PLOS Computational Biology},
     title = {A Kirchhoff-Nernst-Planck framework for modeling large scale extracellular electrodiffusion surrounding morphologically detailed neurons},
     year = {2018},
