+++ .readthedocs.yaml
@@ -27,4 +27,6 @@ sphinx:
 python:
    install:
    - requirements: docs/requirements.txt
+   - method: pip
+      path: .
 
