### 2022-01-21 8ccef6f
Fix title levels in model.md (see #6)
