### 2022-01-21 cc480b8
Fix traversal source when starting from an empty neuron
