### 2022-01-21 aca78cd
Minor improvements of documentation and fix #6
