### 2022-01-22 65b26e4
Add read_swc documentation
