### 2022-01-23 #4012 @cgurrr
Hi @martejulie, we now released version 0.3 and updated the documentation (it is now on readthedocs with continuous updates).
