### 2022-01-24 ngltr/sinaps#4 @martejulie
Please update Figure 1 in the paper accordingly.

I would also like to encourage you to add units to the colorbars.
