### 2022-01-24 5462a2d
Add label to fieldplot colormap (see #6)
