### 2022-01-24 ngltr/sinaps#4 @cgurrr
We have now added the units in the colorbar, and changed the figure in the paper.
