+++ src/sinaps/core/model.py
@@ -10,6 +10,7 @@ Conductance : nS (E-9)
 Capacitance : pF (E-12)
 Current : pA (E-12)
 Mol :aM (E-18)
+Concentration : aM/μm3 = mM/L (E-3)
 
 """
 
