+++ src/sinaps/core/simulation.py
@@ -5,11 +5,12 @@ from functools import reduce
 import numpy as np
 import pandas as pd
 from scipy.integrate import solve_ivp
-from scipy.sparse import csr_matrix, bmat
+from scipy.sparse import csr_matrix, dia_matrix
 from scipy import interpolate
-from scipy.misc import derivative
 from numba import jit
 from tqdm import tqdm
+
+from sinaps.core.species import CHARGE
 
 
 class Simulation:
@@ -52,9 +53,6 @@ class Simulation:
             np.concatenate([np.identity(self.N.nb_comp), self.N._connection_mat()])
         )
         self.G = G @ self.k_c
-        self.v_source, source_mat = self.N._all_Vsource()
-        # if len(self.v_source):
-        #    self.G = bmat([[self.G, source_mat], [source_mat.T, None]]).tocsr()
         self.V_S0 = np.zeros(max(np.concatenate((self.N.idV, self.N.idS))) + 1)
         self.N._fill_V0_array(self.V_S0)
         self.N._fill_S0_array(self.V_S0)
@@ -64,7 +62,15 @@ class Simulation:
         self.sol_diff = dict()
         self.progressbar = progressbar
 
-    def run(self, t_span, method="BDF", atol=1.49012e-8, **kwargs):
+    def run(
+        self,
+        t_span,
+        species=None,
+        temperature=310,
+        method="BDF",
+        atol=1.49012e-8,
+        **kwargs
+    ):
         """Run the voltage related simulation
 
         The results of the simulation are stored in attribute `V`
@@ -77,8 +83,7 @@ class Simulation:
         Other Parameters
         ----------------
         **kwargs :
-            args to pass to the ode solver. see the `scipy solve_ivp
-            <https://docs.scipy.org/doc/scipy/reference/generated/scipy.integrate.solve_ivp.html>`_ doc
+            args to pass to the ode solver. see the `scipy solve_ivp <https://docs.scipy.org/doc/scipy/reference/generated/scipy.integrate.solve_ivp.html>`_ doc
 
         Examples
         --------
@@ -96,24 +101,52 @@ class Simulation:
         else:
             tq = None
 
-        for v in self.v_source:
-            self.V_S0[v.idV] = v.V(t_span[0])
+        if species is None:
+            species = tuple(
+                self.N.species
+            )  # conversion in tuple the ensure the order (N.species i a set)
+
+        reactions = []
+        for reac in self.N.reactions:
+            # Conversion species to int
+            members = [
+                {species.index(sp): n for sp, n in member.items()}
+                for member in reac[0:2]
+            ]
+            if reac[2]:
+                reactions.append((*members, reac[2]))
+            members.reverse()
+            if reac[3]:
+                reactions.append((*members, reac[3]))
+
+        C0 = np.zeros((self.N.nb_comp, len(species)))
+        self.N._fill_C0_array(C0, species)
+        y0 = np.hstack([self.V_S0, np.reshape(C0, -1, "F")])
+        idC = np.array(range(len(self.V_S0), len(y0)))
+        charges = (
+            np.array([CHARGE[s] for s in species]) * 96.48533132838746
+        )  # pA/aMol = C/mol/ms E-3
 
         sol = solve_ivp(
-            lambda t, y: Simulation._ode_function(
+            lambda t, y: self._ode_function(
                 y,
                 t,
                 self.idV,
                 self.idS,
+                idC,
                 self.Cm1,
                 self.G,
-                self.v_source,
                 self.channels.values(),
+                species,
[…]
