### 2022-05-05 ngltr/sinaps#14 @meg-simula
Requirements specification missing
A specification of the sinaps requirements seems to be missing. Could you please include. (For instance, I'm unable to test functionality due to missing numba.experimental package it seems.)
