### 2022-05-05 4573fe4
Update numba minimum version
