### 2022-05-05 ngltr/sinaps#14 @ngltr
I think it is because you have an old version of numba (<0.49) already installed,
Can you try
`pip install numba --upgrade` ?
We will also add a minimum version for numba in the requirements specification
