### 2022-05-08 #4012 @cgurrr
@meg-simula, see https://github.com/ngltr/sinaps/issues/14#issuecomment-1119000622
It seems that you need to upgrade numba. We now have added a minimum version requirement in the install dependencies. You could also simply update sinaps to 0.3.1., and it should work.
Let us now if you have any more issues.
