### 2022-05-10 #4012 @meg-simula
Thanks @cgurrr for the quick response and updated dependencies instructions. I am still unable to test the sinaps package, now due to errors in hvplot:
