### 2022-05-10 #4012 @ngltr
@meg-simula, The installation process works well from scratch ( for example in the automated test, dependencies installation is checked https://github.com/ngltr/sinaps/runs/6341186525?check_suite_focus=true), but it seems we need to add other minimum version requirements for the all dependencies in case the user have old version already installed.
As a temporary solution, can you try to upgrade all dependencies :
`pip install  hvplot pandas networkx  datashader numpy --upgrade`
If sill not working maybe, you can try to install sinaps in a virtualenv :
`python3 -m venv sinaps_env`
`source sinaps_env/bin/activate`
`pip install sinaps`
