### 2022-05-10 #4012 @meg-simula
@cgurrr Would you please take another look at the references in the paper? Last names seems to have gone missing mostly, with only initials remaining.

After this,  could you make a tagged release and archive for the software, and report the version number and archive DOI here?
