### 2022-05-12 f081171
Update installing.rst
