### 2022-05-12 #4012 @cgurrr
@meg-simula Many thanks for your comments.
- We added a note on potential incompatibilities and  proposed potential fixes in the documentation: https://sinaps.readthedocs.io/en/stable/installing.html
- We corrected the author name syntax in the references of the paper.
- The tagged release is now 0.3.2,
- Finally, we made an archive in Zenodo. DOI:  `10.5281/zenodo.6540844`
